completo vistas de mensajeria, y agrego confirmacion de borrado a blog y mensajeria

+++ blog/views.py
@@ -71,7 +71,6 @@ def nuevaEntrada(request):
             entrada.save()
             return redirect('blog:inicioUsuario')
         else:
-            print(entradaForm.errors)
             messages.error(request,"Los datos ingresados no son válidos")
             return render(request, 'album/nuevaEntrada.html')
     else:
@@ -83,12 +82,14 @@ def nuevaEntrada(request):
 def eliminarEntrada(request, entr_id):
     ''' eliminar entrada (debe ser el usuario que la creó) '''
     entrada = Entrada.objects.get(id=entr_id)
-    titulo = entrada.titulo
-    entrada.delete()
-    print(entr_id)    
-    print(entrada)
-    messages.success(request, f"Entrada '{titulo}' eliminada")
-    return redirect('blog:inicioUsuario')
+    if request.method == 'POST':    
+        entrada = Entrada.objects.get(id=entr_id)
+        titulo = entrada.titulo
+        entrada.delete()
+        messages.success(request, f"Entrada '{titulo}' eliminada")
+        return redirect('blog:inicioUsuario')
+    else:
+        return render(request, 'album/confirmarBorrado.html', {'entrada':entrada})
 
 
 @login_required
@@ -106,7 +107,6 @@ def editarEntrada(request, entr_id):
             entrada.save()
             return redirect('blog:inicioUsuario')
         else:
-            print(entradaForm.errors)
             messages.success(request,"Los datos no son correctos")
             return render(request, 'album/editarEntrada.html')
     else:
@@ -125,7 +125,7 @@ def mostrarEntrada(request, entr_id):
     entrada = Entrada.objects.get(id=entr_id)
     return render(request, 'album/mostrarEntrada.html', {"entrada":entrada})
 
-@login_required(login_url='login')
+@login_required(login_url='blog:login')
 def likeEntrada(request, entr_id):
     entrada=Entrada.objects.get(id=entr_id)
     entrada.likes = entrada.likes + 1
